Add MemoryCard detection as latest sav, pkmdb

+++ PKHeX.Core/Saves/Util/Recognition/SaveHandlerARDS.cs
@@ -10,7 +10,7 @@ public sealed class SaveHandlerARDS : ISaveHandler
     private const int sizeHeader = 0xA4;
     private const int ExpectedSize = SaveUtil.SIZE_G4RAW + sizeHeader; // 0x800A4
 
-    public bool IsRecognized(int size) => size is ExpectedSize;
+    public bool IsRecognized(long size) => size is ExpectedSize;
 
     public SaveHandlerSplitResult TrySplit(ReadOnlySpan<byte> input)
     {
